@@ -1,63 +1,77 @@
-﻿// See https://aka.ms/new-console-template for more information
+﻿using System;
+using System.IO;
+using Cocona;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Hosting;
 using Serilog;
-using System;
-using System.IO;
-using LayerComparerConsole;
+using Serilog.Extensions.Logging;
+using Serilog.Sinks.SystemConsole;
+using Serilog.Enrichers;
+using Serilog.Sinks.File;
+using Serilog.Settings.Configuration;
 
-namespace OGCGeometryValidatorCore
+namespace LayerComparerConsole;
+
+/// <summary>
+/// template using DI, Serilog, Settings 
+/// </summary>
+class Program
 {
-    /// <summary>
-    /// template using DI, Serilog, Settings 
-    /// </summary>
-    class Program
+    static void Main(string[] args)
     {
-        static void Main(string[] args)
+        // config logger first
+        var configuration = new ConfigurationBuilder()
+            .SetBasePath(Directory.GetCurrentDirectory())
+            .AddJsonFile("appsettings.json", optional: false, reloadOnChange: true)
+            .Build();
+
+
+        Log.Logger = new LoggerConfiguration()
+            .ReadFrom.Configuration(configuration)
+            .WriteTo.Console()
+            .CreateLogger();
+
+        Log.Logger.Information("Application starting");
+
+
+        var builder = CoconaApp.CreateBuilder();
+
+        builder.Host.UseSerilog();
+
+        builder.Services.AddTransient<ILayerCompareService, LayerCompareService>();
+
+        var app = builder.Build();
+
+        // Definieren Sie einen Command, der ILayerCompareService verwendet
+        app.AddCommand("compare", (
+            [Argument(Description = @"path to master GDB, e.g. G:\BnL\Daten\Ablage\DNL\Bundesinventare\Auengebiete\Auengebiete.gdb")]
+            string masterGdbPath,
+            [Argument(Description = "name of master layer to compare, e.g. N2017_Revision_Auengebiete_Anhang2_20171101")]
+            string masterLayer,
+            [Argument(Description = @"path to candidate GDB, e.g. G:\BnL\Daten\Ablage\DNL\Bundesinventare\Auengebiete\vonBIOPzurValidierung20210219\AU.gdb")]
+            string candidateGdbPath,
+            [Argument(Description = "name of candidate layer to compare, e.g. au_Anhang2_20171101")]
+            string candidateLayer,
+            ILayerCompareService layerCompareService) => // DI für ILayerCompareService
         {
-            var builder = new ConfigurationBuilder();
-            BuildConfig(builder);
 
-            // config logger first
-            Log.Logger = new LoggerConfiguration()
-                .ReadFrom.Configuration(builder.Build())
-                .Enrich.FromLogContext()
-                .WriteTo.Console()
-                .CreateLogger();
+            Console.WriteLine($"Running.. ");
+            // Verwenden Sie die Argumente und den Service, um die Logik auszuführen
+            //layerCompareService.Compare(args);
+            layerCompareService.Compare(new string[] { masterGdbPath, masterLayer, candidateGdbPath, candidateLayer });
 
-            Log.Logger.Information("Application starting");
+            Console.WriteLine("Vergleich durchgeführt.");
 
-            // set up DI
-            var host = Host.CreateDefaultBuilder()
-                .ConfigureServices((context, services) =>
-                {
-                    services.AddTransient<ILayerCompareService, LayerCompareService>();
-                })
-                .UseSerilog()
-                .Build();
+            Console.WriteLine();
+            Console.WriteLine("Press ENTER to end..");
+            Console.ReadLine();
 
-            // start the service in the host
-            var svc = ActivatorUtilities.CreateInstance<LayerCompareService>(host.Services);
-            svc.Run(args);
-        }
 
-        /// <summary>
-        /// Set up logging manually.
-        /// CreateTest connection to appsettings manually with logging configuration
-        /// </summary>
-        /// <param name="builder"></param>
-        static void BuildConfig(IConfigurationBuilder builder)
-        {
-            // get overwriteable settings, first appsetting, then - if available - .Development. or .Production.json,
-            // that will overwrite settings from appsettings (like CSS)
-            builder.SetBasePath(Directory.GetCurrentDirectory())
-                .AddJsonFile("appsettings.json", optional: false, reloadOnChange: true)
-                .AddJsonFile(
-                    $"appsettings.{Environment.GetEnvironmentVariable("ASPNETCORE_ENVIRONMENT") ?? "Production"}.json",
-                    optional: true)
-                .AddEnvironmentVariables(); // eg. Default-connection string to sql express in the local env variables
-        }
+        });
 
+        app.Run();
     }
 }
+
+
